move clipping code to its own file
