+++ include/ramiel/graphics.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "../src/graphics/camera.h"
+#include "../src/graphics/clip.h"
 #include "../src/graphics/effects.h"
 #include "../src/graphics/entity.h"
 #include "../src/graphics/light.h"
